+++ packages/@jsii/python-runtime/src/jsii/_kernel/__init__.py
@@ -50,7 +50,6 @@ from .types import (
 )
 from .._utils import Singleton
 
-
 _nothing = object()