+++ packages/@jsii/python-runtime/src/jsii/_kernel/providers/__init__.py
@@ -1,5 +1,4 @@
 from .base import BaseProvider
 from .process import ProcessProvider
 
-
 __all__ = ["BaseProvider", "ProcessProvider"]